ml_backend/app.py
import os
import sys
from flask import Flask, jsonify
from flask_cors import CORS
from dotenv import load_dotenv

load_dotenv()

current_dir = os.path.dirname(os.path.abspath(__file__))
sys.path.append(current_dir)

from routes.similar_case_route import similar_case_bp

app = Flask(__name__)

app.register_blueprint(similar_case_bp, url_prefix='/find-similar-cases')

@app.route('/health', methods=['POST'])
def health_check():
    return jsonify({"status": "healthy"}), 200

if __name__ == '__main__':
    port = int(os.getenv("PORT", 5001))
    app.run(host='0.0.0.0', port=port, debug=True)