ml_backend/requirements.txt
Flask
Flask-Cors
python-dotenv
pymongo
google-generativeai
numpy
scikit-learn
nltk
setuptools
sentence_transformers